# Manually vectorize distance functions to improve performance