--- a/src/annoylib.h
+++ b/src/annoylib.h
@@ -613,6 +613,32 @@ struct Manhattan : Minkowski{
       d += fabs((*x) - (*y));
     return d;
   }
+#ifdef USE_AVX
+  static inline float distance(const float* x, const float* y, int f) {
+    float result = 0;
+    int i = f;
+    if (f > 7) {
+      __m256 manhattan = _mm256_setzero_ps();
+      __m256 minus_zero = _mm256_set1_ps(-0.0f);
+      for (; i > 7; i -= 8) {
+        const __m256 x_minus_y = _mm256_sub_ps(_mm256_loadu_ps(x), _mm256_loadu_ps(y));
+        const __m256 distance = _mm256_andnot_ps(minus_zero, x_minus_y); // Absolute value of x_minus_y (forces sign bit to zero)
+        manhattan = _mm256_add_ps(manhattan, distance);
+        x += 8;
+        y += 8;
+      }
+      // Sum all floats in manhattan register.
+      result = hsum256_ps_avx(manhattan);
+    }
+    // Don't forget the remaining values.
+    for (; i > 0; i--) {
+      result += fabsf(*x - *y);
+      x++;
+      y++;
+    }
+    return result;
+  }
+#endif
   template<typename S, typename T, typename Random>
   static inline void create_split(const vector<Node<S, T>*>& nodes, int f, Random& random, Node<S, T>* n) {
     vector<T> best_iv(f, 0), best_jv(f, 0);