For a start I vectorized the euclidian distance with AVX intrinsics. I wanted to make sure, that there is an actual improvement. I created a super [basic benchmark](https://gist.github.com/ReneHollander/4ca884ceb1537f013aa8d0f868e9b844) which improved the index build time with `2*f` trees from roughly 45 seconds to 35 seconds and search time from 13ms to 9ms (Windows 10 Subsystem for Linux on an i7 4770K overclocked to 4GHz, the benchmark also contains some results). I only tested with random high dimensional data (96 and 100 dimensions and 100000 features)

With that said, is there interest in optimizing other distance functions and other functions in general? (I am no expert and there is the possibility, that there are even better solutions, but for a start it could help)