@erikbern 2018-01-14
that's pretty cool! I think the small differences in the distances is totally acceptable (after all, the A in Annoy stands for Approximate)

I might run some tests and merge this if I can reproduce the speedup!