@a1k0n 2018-01-14
Would be good to try it on the dot product (for cosine distance, and for every single plane test when traversing the tree) also, if it's not already auto-vectorized...