@erikbern 2018-01-15
Good point @a1k0n – should add that too and test!