@ReneHollander 2018-01-15
Sure! I will try to vectorize some more functions (first with AVX) and later with SSE. Sadly I am not able to look into AVX512, because my CPU doesn't support it.