@ReneHollander 2018-01-16
After some testing I found an issue in the way features with a dimension that is not a multiple of 8 are handled. After fixing those calculations, all tests pass now.
I also optimized the dot product in the margin function and the angular distance. Here are the results: https://gist.github.com/ReneHollander/4def83f7755c6d63e1efb120774f2306
The speedups for index building are quite dramatic, but also the query time got reduced quite a bit.
I did some profiling with valgrind and also found a hotspot with the inner loops in `two_means`. I don't think optimizing it will improve the speed as much as my other changes did, but it could help. Some goes with the vector norm.