@ReneHollander 2018-01-16
I also forgot to add, that those optimizations are enabled if AVX support is found (building with `-march=native`, as it is currently done in `setup.py`, will pick it up). One can override the hand vectorization with defining `NO_MANUAL_VECTORIZATION` before including `annoylib.h`.

I also don't know if it is of use adding support for only SSE CPUs. Every modern CPU since 2011 has support for AVX.