@erikbern 2018-01-16
excellent – will take a look at this in the next few days