@ReneHollander 2018-01-16
`get_norm`, `normalize` and `two_means` are now also optimized with AVX intrinsics. This only speeds up trees using angular distance ([Benchmark](https://gist.github.com/ReneHollander/98f8a231fc1725fc2e3e388bff5ed741), baseline is with hand optimized distance and margin), euclidian distance is completly unaffected, as expected.

There are also intrinsics for popcount, so it should be possible to optimize hamming distance by hand as well. Also manhattan distance is missing the optimizations for `distance` (I have to look into the absolute value calculation using AVX first).