@ReneHollander 2018-01-17
I decided against vectorizing hamming distance calculation. Because the data is packed into 64bit integers, it would only make sense when you have features with over 256 dimensions.
Other than that, everything is finished and it just needs to be reviewed 😄