@erikbern 2018-01-17
This is great – let me take a look soon!