@erikbern 2018-01-19
Thanks @ReneHollander – this is great!

I think the code is a bit repetitive so let me refactor and share a PR shortly!