@ReneHollander 2018-01-19
You are right, the code is pretty repetitive. Maybe it is possible to move out the dot product calculation. I also fixed some potential windows incompatibility (and the unused code) in another pull request: #266