@erikbern 2018-01-19
@ReneHollander did you see #265 ? I factored out two helpers

* `dot(p, q)` for dot product
* `dot3(p, q)` for computing the quantities pq, pp, and qq